Moved all tests to common class
